PRO-3554: Refactored submit service to account for movement of registry code to the persistence service
