+++ src/lib/AI-Agent.xml
@@ -52,13 +52,6 @@
             <Method name="submit" reportExecutionTime="true"/>
             <Method name="resubmit" reportExecutionTime="true"/>
         </Class>
-        <Class name="uk.gov.hmcts.probate.services.submit.services.SequenceService">
-            <Method name="nextRegistry" reportExecutionTime="true"/>
-            <Method name="populateRegistrySubmitData" reportExecutionTime="true"/>
-            <Method name="populateRegistryResubmitData" reportExecutionTime="true"/>
-            <Method name="getRegistrySequenceNumber" reportExecutionTime="true"/>
-            <Method name="identifyNextRegistry" reportExecutionTime="true"/>
-        </Class>
         <Class name="uk.gov.hmcts.probate.services.submit.services.SubmitService">
             <Method name="submit" reportCaughtExceptions="true" reportExecutionTime="true"/>
             <Method name="resubmit" reportCaughtExceptions="true" reportExecutionTime="true"/>
