+++ src/main/java/uk/gov/hmcts/probate/services/submit/clients/PersistenceClient.java
@@ -21,8 +21,8 @@ public class PersistenceClient {
     @Value("${services.persistence.submissions.url}")
     private String submissionsPersistenceUrl;
 
-    @Value("${services.persistence.sequenceNumber.url}")
-    private String sequenceNumberPersistenceUrl;
+    @Value("${services.persistence.registry.url}")
+    private String registryPersistenceUrl;
 
     private RestTemplate restTemplate;
     private PersistenceEntityBuilder builder;
@@ -69,8 +69,8 @@ public class PersistenceClient {
     }
 
     @Retryable(backoff = @Backoff(delay = 100, maxDelay = 500))
-    public Long getNextSequenceNumber(String registryName){
-        ResponseEntity<Long> response = restTemplate.getForEntity(sequenceNumberPersistenceUrl + "/" + registryName, Long.class);
+    public JsonNode getNextRegistry(long submissionReference){
+        ResponseEntity<JsonNode> response = restTemplate.getForEntity(registryPersistenceUrl + "/" + submissionReference, JsonNode.class);
         return response.getBody();
     }
 }
