+++ src/main/java/uk/gov/hmcts/probate/services/submit/health/SubmitHealthConfiguration.java
@@ -9,10 +9,10 @@ import org.springframework.web.client.RestTemplate;
 
 @Configuration
 public class SubmitHealthConfiguration
-{	
+{
     @Autowired
     private RestTemplate restTemplate;
-    
+
     @Value("${services.coreCaseData.baseUrl}")
     private String servicesCcdBaseUrl;
 
@@ -25,12 +25,12 @@ public class SubmitHealthConfiguration
     @Bean
     @ConditionalOnProperty(prefix = "services.coreCaseData", name = "enabled", matchIfMissing = true)
     public SubmitHealthIndicator ccdServiceHealthIndicator() {
-    	return new SubmitHealthIndicator(servicesCcdBaseUrl, restTemplate); 
-    }  
-    
+    	return new SubmitHealthIndicator(servicesCcdBaseUrl, restTemplate);
+    }
+
     @Bean
     public SubmitHealthIndicator persistenceServiceHealthIndicator() {
-    	return new SubmitHealthIndicator(servicesPersistenceBaseUrl, restTemplate); 
+    	return new SubmitHealthIndicator(servicesPersistenceBaseUrl, restTemplate);
     }
 
     @Bean
