+++ src/main/resources/application.yml
@@ -48,7 +48,7 @@
       baseUrl: http://localhost:8282
       formdata.url:  ${services.persistence.baseUrl}/formdata
       submissions.url:  ${services.persistence.baseUrl}/submissions
-      sequenceNumber.url:  ${services.persistence.baseUrl}/sequence-number
+      registry.url:  ${services.persistence.baseUrl}/registry
 
     coreCaseData:
       baseUrl: https://case-data-app.dev.ccd.reform.hmcts.net:4451
@@ -136,29 +136,3 @@
 
   azure:
     app_insights_key: ${APPINSIGHTS_INSTRUMENTATIONKEY:false}
-
-  registries:
-    - name: oxford
-      id: 0
-      email: oxford@email.com
-      address: |
-        Line 1 Ox
-        Line 2 Ox
-        Line 3 Ox
-        PostCode Ox
-    - name: birmingham
-      id: 1
-      email: birmingham@email.com
-      address: |
-        Line 1 Bham
-        Line 2 Bham
-        Line 3 Bham
-        PostCode Bham
-    - name: manchester
-      id: 2
-      email: manchester@email.com
-      address: |
-        Line 1 man
-        Line 2 man
-        Line 3 man
-        PostCode Manchester
